OnShow/OnHide hooks for Forms.
These should probably be implemented better, perhaps using
connect_signal, _emit and a proper signals dictionary, but I'm not too
certain that this works best.

It does provide a simple way of unhooking CallEvery functions when the
form that it is linked to is hidden or destroyed, or at the very least
pausing them.