A short function to return a reference to the previous form.
It neither shows the previous form nor discards it from the list of
previous forms. This function can be useful if you want to adjust
something on a previous form. Say you've added a new item to a list and
saved it: you may wish to update the list editor in the previous to
reflect the new item.