Provide a "focus" specifier for ButtonDialog, fix width.
The function will now compare the value of the focus parameter with the
button label (both lowercased) when building the button list. If it
matches, it will specify that this button should be given initial focus,
rather than whichever button is the first.

This also provides an urwid.Padding() wrapper around each widget,
preventing them from taking up the entirety of the dialog.